[ATL_APITEST] Add 'atltest.h' and use it
