+++ modules/rostests/apitests/atl/CAtlList.cpp
@@ -9,27 +9,7 @@
 #ifdef HAVE_APITEST
     #include <apitest.h>
 #else
-    #include <stdlib.h>
-    #include <stdio.h>
-    #include <stdarg.h>
-    int g_tests_executed = 0;
-    int g_tests_failed = 0;
-    void ok_func(const char *file, int line, bool value, const char *fmt, ...)
-    {
-        va_list va;
-        va_start(va, fmt);
-        if (!value)
-        {
-            printf("%s (%d): ", file, line);
-            vprintf(fmt, va);
-            g_tests_failed++;
-        }
-        g_tests_executed++;
-        va_end(va);
-    }
-    #undef ok
-    #define ok(value, ...)  ok_func(__FILE__, __LINE__, value, __VA_ARGS__)
-    #define START_TEST(x)   int main(void)
+    #include "atltest.h"
 #endif
 
 #include <atlbase.h>
@@ -73,9 +53,4 @@ START_TEST(CAtlList)
         index++;
     }
     ok(it == NULL, "it does still point to something!\n");
-
-#ifndef HAVE_APITEST
-    printf("CAtlList: %i tests executed (0 marked as todo, %i failures), 0 skipped.\n", g_tests_executed, g_tests_failed);
-    return g_tests_failed;
-#endif
 }
