+++ modules/rostests/apitests/atl/CSimpleMap.cpp
@@ -8,27 +8,7 @@
 #ifdef HAVE_APITEST
     #include <apitest.h>
 #else
-    #include <stdlib.h>
-    #include <stdio.h>
-    #include <stdarg.h>
-    int g_tests_executed = 0;
-    int g_tests_failed = 0;
-    void ok_func(const char *file, int line, bool value, const char *fmt, ...)
-    {
-        va_list va;
-        va_start(va, fmt);
-        if (!value)
-        {
-            printf("%s (%d): ", file, line);
-            vprintf(fmt, va);
-            g_tests_failed++;
-        }
-        g_tests_executed++;
-        va_end(va);
-    }
-    #undef ok
-    #define ok(value, ...)  ok_func(__FILE__, __LINE__, value, __VA_ARGS__)
-    #define START_TEST(x)   int main(void)
+    #include "atltest.h"
 #endif
 
 #include <atlbase.h>
@@ -198,9 +178,4 @@ START_TEST(CSimpleMap)
     ok(map1.GetSize() == 1, "Expected map1's size is 1, was %d\n", map1.GetSize());
     ok(!!map1.RemoveAt(0), "Expected RemoveAt(0) to succeed\n");
     ok(map1.GetSize() == 0, "Expected map1's size is 0, was %d\n", map1.GetSize());
-
-#ifndef HAVE_APITEST
-    printf("CSimpleMap: %i tests executed (0 marked as todo, %i failures), 0 skipped.\n", g_tests_executed, g_tests_failed);
-    return g_tests_failed;
-#endif
 }
